refactor: split lib_tests.rs into module-specific test files
Distribute 150 tests from the monolithic lib_tests.rs to match the
module structure created in the prior split. No test behavior changes.

- column_matching_tests.rs (80 tests)
- response_tests.rs (51 tests)
- request_tests.rs (17 tests)
- lib_tests.rs reduced to 2 cross-cutting default tests
